requirements.txt
python-telegram-bot==20.7
selenium==4.20.0